restructure icon components
