+++ apps/browser/src/autofill/content/components/constants/styles.ts
@@ -161,6 +161,21 @@ export const typography = {
   `,
 };
 
+export const ruleNames = {
+  fill: "fill",
+  stroke: "stroke",
+} as const;
+
+type RuleName = (typeof ruleNames)[keyof typeof ruleNames];
+
+/*
+ * `color` is an intentionally generic name here, since either fill or stroke may apply, due to
+ * inconsistent SVG construction. This consequently precludes dynamic multi-colored icons here.
+ */
+export const getIconColorRule = (color: string, rule: RuleName = ruleNames.fill) => `
+  ${rule}: ${color};
+`;
+
 export function scrollbarStyles(theme: Theme) {
   return {
     default: `
