+++ apps/browser/src/autofill/content/components/icons/index.ts
@@ -1,0 +1,12 @@
+export { AngleDown } from "./angle-down";
+export { BrandIconContainer } from "./brand-icon-container";
+export { Business } from "./business";
+export { Close } from "./close";
+export { ExclamationTriangle } from "./exclamation-triangle";
+export { Family } from "./family";
+export { Folder } from "./folder";
+export { Globe } from "./globe";
+export { PartyHorn } from "./party-horn";
+export { PencilSquare } from "./pencil-square";
+export { Shield } from "./shield";
+export { User } from "./user";
